+++ lib/flutter_post_frame.dart
@@ -1,6 +1,9 @@
 import 'dart:async';
 import 'package:flutter/widgets.dart';
 
+export 'post_frame_builder.dart';
+export 'post_frame_scroll_builder.dart';
+
 mixin FlutterPostFrameMixin<T extends StatefulWidget> on State<T> {
   @override
   void initState() {